refactor: extract Stripe webhook handlers into modular utility functions
